+++ Gruntfile.js
@@ -92,8 +92,8 @@ module.exports = function(grunt) {
 						'assets/themes/WPadmin/css/colors.css',
 						'assets/themes/WPadmin/css/font.css',
 						'assets/themes/WPadmin/css/stylesheet.css',
-						'assets/themes/WPadmin/css/rtl.css',
-						'assets/themes/WPadmin/css/zresponsive.css'
+						'assets/themes/WPadmin/css/zresponsive.css',
+						'assets/themes/WPadmin/css/rtl.css'
 					],
 					'assets/themes/WPadmin/stylesheet_wkhtmltopdf.css': [
 						'assets/themes/WPadmin/css/colors.css',